Implementing parallax backgrounds
Reworked the default html to have 2 background images that move at different speeds as page is scrolled. The .wrap class now exists to support the parallax effect, it has the initial perspective and transform-style declarations that make the transform declarations on the images work.
NOTE: the images DO NOT enclose the content, although .wrap does. Using non-repeating images is a little trickier: each image container has to be sized big enough to cover the viewport when scrolling, but it shouldn't push content down. So use same height and NEGATIVE bottom margin of at least 100vh. Width works at same multiple of vw.